Remove the last of the compiler warnings.
Thanks for all the wonderful work that has been done on this library. I figured that the least I could do would be to spend a few minutes and help improve it. Here are my suggested changes:

Since processRTCM will only be used if user-defined, I suggest that it be commented out completely. 
Changed the literal integer constant to unsigned to remove the related warning. 
Remove the unused maxWait from sendI2cCommand.
If getGeoidSeparation is not only redundant but has no associated code, it may be better to remove it to avoid the maxWait warning.