Add user context to library
The current overwrite method of weak pointer is a global overwrite. As class
is not polymorphic type (aka the overwrite enabled methods are weak instead of
virtual) and there is no vtable as none of the methods are virtual, there is
no way to have instance specific overwrite.

A user context will provide a method to call back into a context of a parent
class and resolve the issue without making the class polymorphic.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>